requirements.txt
Brotli==1.0.9
click==8.1.3
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.12
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
plotly==5.10.0
python-dateutil==2.8.2
pytz==2022.2.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
Werkzeug==2.2.2
xgboost==1.6.2
tensorflow